+++ error.go
@@ -1,0 +1,7 @@
+package engine
+
+import "errors"
+
+var (
+	ErrCanceled = errors.New("the order has been canceled")
+)