+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # CHANGELOG
 
+## v0.3 (2022-06-05)
+
+- add order book update event
+- add time in force strategy (gtc, ioc)
+- add "post only" behavior
+
 ## v0.2 (2022-05-11)
 
 - use skiplist algorithm for order queue